Wrestling with issue 2! Variable names!
https://github.com/BlitzerSIO/grass-cheetah/issues/2

Pinning down potential causes, but not really sure if I've identified the
REAL reason why it's happening. Still, it mostly works as-is right now. It
will need fixing all the same though.

Also, doing some variable renaming to be more readable both in terms of
consistency and in a "brief glance" of what the variable does.
* GSP is now ground_speed
* Similarly, gsp_vector is now ground_speed_vector
* frc is friction (the other use of friction in the code as a local
  variable name is now called ground_friction)
* acc and dec are now accel and decel
* top is top_speed
* slp is slope
* spec is special_number

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>