Fixed issue 2? Kind of.
Issue 2 *might* be fixed, but it's a hacky fix I'm not really happy about
and would like the other devs to take a look at please.

Even if the main part of issue 2 is fixed, https://github.com/BlitzerSIO/grass-cheetah/issues/2#issuecomment-416033663
is still an issue. I suspect it's not really that closely related to the
other problems though, so might open up a separate issue for that one.

Also, got rid of SonicCamera.gd as it does nothing.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>