Collision shape woes.
Noticed the way that the collision shape for the Sonic sprite had been
made was incorrect, so modified that and also brought the collision rays in
a little. Don't think it fixes anything, but it doesn't seem to have made
anything worse, either. So shall keep an eye on it. Wall detection needs
fixing, in any case.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>