HUD works, part one.
Got the HUD up and working, via getters and setters. Cannot test anything
but score yet as need to add a timer for the time to work, and no rings or
lives functionality yet exposed to the getters/setters.

This is using the game_space singleton from my pre-existing Sonic engine
project.

This also paves the way for game over/time over to be added to the game.
Added sounds for getting/losing rings to the game.

Also renamed a few more variables to be clearer about what they do - and
found myself wondering about why the is_on_floor function is not being
used, shall try that and see what happens as it seems it'd possibly be
simpler than the way it's currently being handled.

Badniks can take a variable number of hits before being destroyed, and have
a score value attached to them.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>