More genericness and print fixes.
game_space has two variables - player_character and player_character_node -
the first is the absolute path to the player character, the second is the
value/reference of the node proper. The former is more for conditions like
has_node; the latter is probably the one going to be used more in other
checks, assignments and the like. Ultimately this is done so that, if the
game does allow for different playable characters, it shouldn't be (too)
difficult to select one.

Functions/nodes changed to use these appear to work. Hopefully there are no
hidden bugs, but we shall see.

Also, got rid of all "print" statements and replaced them with "printerr".
Not that the player should see any of them anyway...

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>