Death works... kinda maybe sorta!
As in, it appears to work, but nothing else to do with it does. Needs to
reset the HUD and all sorts of things. But the (very) bare basics work.

You can "test" this "feature" by pressing CTRL+F1 in game. Be warned, it
needs a *lot* of work to make it actually usable!

This is the last commit for Sslaxxworks - after this it is going to be
merged into master.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>