Some state changes, other things.
For one, IDLE_STATE should be 0. The way it'd been defined before it was 1.
IDLE_STATE is the absence of state; nothing is going on. Starting to add
checks for CUTSCENE_STATE, too; this state should not prevent movement etc;
it should only prevent the player controlling the character. Going to be
adding checks for movement states before setting IDLE_STATE; not doing this
seems to be the reason for the player character being "slidy" (because of
being in the idle animation) when coming to a stop.

Renamed some of the functions for consistency's sake.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>