Starting work on a title screen!
[DON'T MERGE THIS INTO MASTER UNTIL AFTER THE "player_generic.gd(changes)"
BRANCH HAS BEEN.]

Starting work on a title screen, trying to get the feel for it. It's using
the background from Crimson District Zone (as that's all there is currently
working). This is just a mock, trying to get a feel, so nothing serious in
here yet. But looking at the possibility (when more zones are done) of
having the background change periodically to that of other zones.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>