Attempting to use the Godot Sonic Engine...
...and the result isn't great. Terrible? Not quite, but very far from good
either.

The "physics" are very far removed from acceptable Sonic type physics.

The slopes and ramps are not being stuck to.

Acceleration to allow access to the upper parts of the level is not there
at all. This is probably due to the physics implementation's way of using
gravity.

The CDZ files have been shifted over to the Godot Sonic Engine's layout. It
displays perfectly, no problems there.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>