Last 3.0 commit for this branch.
This branch is being switched to 3.1 in preparation for the final release
(hopefully) at the end of this month.

This means that, after this commit, IT WILL NOT OPEN IN 3.0. For the
purposes of this branch, until 3.1 stable is released this branch is going
to use the beta1 release of 3.1.

Going to take some of the other branch's code and apply it here. The
raycasts should *only* be used for floor edge detection (i.e., what
animation to play). So they should only be activated when `is_on_floor` is
true. Snapping to the ground is going to be enabled via the new
`move_and_slide_with_snap` function; hopefully it should at least remove
some of the issues with this branch's sticking to the floor. The new
`sync_to_physics` boolean should be looked at too.

Fixes to the acceleration/deceleration code, hopefully shall make them
more consistent.

Gravity implementation is still off, so that needs looking at more. Same
with jumping; going to adapt the other branch's jumping code for this one.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>