Minor stuff and fixes.
Removed some of the stuff that doesn't really belong with this repo.

Started on the rings. Going to need to edit the sprite asset for them.

Adjusted jump height a little more (though the way that's working right now
is going to change a bit). Also changed some logic in getters and setters.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>