Misadventures with pause!
At the moment, trying to get pause to work. It's not doing what I'd expect
it to do, so at the mo it's disabled. Hopefully (he says!) it won't be too
much work to get it to behave.

What it was doing was pausing the game, but not displaying the popup, nor
was unpausing the game resuming anything.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>