Pivot and edge detection started...
Not doing anything yet, but the RayCast2Ds for floor edge detection and
pivoting the player character to the slope of the surface will be enabled
or disabled when the player is on the ground or in the air.

Also there're some of the Crimson District assets ready to be re-added to
the level when the time comes.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>