Checkpoints and "air friction".
Air friction isn't, really, but it seems to do the job sufficiently. Might
need fine tuning down the road, but right now it will do. Added a few extra
changes for helping with floor snapping. Also sanity checking for the
speed helper functions.

Started to add checkpoint functionality in to the game.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>