Rotation!
Rotation is now active. So far, it seems to work mostly as expected, too.
There is some sanity checking in there, seems to do the job.

Might want to look at the debug HUD and how to change it, as it's having to
report so much info now it's getting rather cluttered looking.

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>