Physics, I hate you.
Lots of sanity checking is likely to be needed before more of the issues
that are happening entirely at random (seemingly) can be fixed. There may
be Godot bugs involved.

https://github.com/godotengine/godot/issues/25759

Prevented super-powered jumping (I think) that can sometimes happen by
jumping when running up a slope (or if velocity.y is non-zero).

Signed-off-by: Stuart "Sslaxx" Moore <stuart@sslaxx.co.uk>